Refactor code for better readability
