+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "paifulogger/src/mjlog2mjai"]
-	path = paifulogger/src/mjlog2mjai
+[submodule "paifulogger/mjlog2mjai"]
+	path = paifulogger/mjlog2mjai
 	url = https://github.com/Jim137/mjlog2mjai.git
