+++ paifulogger/__init__.py
@@ -12,10 +12,10 @@ del os
 from .version import __version__
 # ------------------------------
 
+from .i18n import localized_str
 from .log import config_path, get_log_func, log, log_paifu, log_parser, remove_old_paifu
 from .paifu_dl import paifu_dl
-from .src.get_paifu import get_paifu, get_paifu_from_client_log, get_paifu_from_local
-from .src.i18n import localized_str
+from .utils import get_paifu, get_paifu_from_client_log, get_paifu_from_local
 
 
 __all__ = [
