+++ paifulogger/logfile/log_into_csv.py
@@ -4,8 +4,8 @@ import re
 
 import pandas as pd
 
-from .i18n import LocalStr
-from .Paifu import Paifu
+from ..i18n import LocalStr
+from ..paifu import Paifu
 
 
 def log_into_csv(paifu: Paifu, local_lang: LocalStr, output: str):
