+++ paifulogger/logfile/log_into_template.py
@@ -1,5 +1,5 @@
-from .i18n import LocalStr
-from .Paifu import Paifu
+from ..i18n import LocalStr
+from ..paifu import Paifu
 
 def log_into_template(paifu: Paifu, local_lang: LocalStr, output: str):
     raise NotImplementedError
