+++ paifulogger/paifu/__init__.py
@@ -1,0 +1,1 @@
+from .paifu import Paifu
