+++ paifulogger/paifu/paifu.pyi
